Create README.md
# Python Game and Application Program

This project is a collection of Python programs that I have created for games and applications. The games are simple and easy to play, while the applications are useful and practical.

## Games

* **Tic-Tac-Toe:** A classic tic-tac-toe game that you can play against the computer. The game is played on a 3x3 grid, and each player takes turns placing their symbol (X or O) in a square. The first player to get three of their symbols in a row (horizontally, vertically, or diagonally) wins the game.
* **Snake:** A classic snake game that you can play to improve your reflexes. The game is played on a grid, and the player controls a snake that is always moving forward. The snake grows longer as it eats food, but it dies if it runs into itself or the edge of the grid. The goal of the game is to get the snake as long as possible without dying.
* **Flappy Bird:** A challenging game where you have to fly a bird through pipes without hitting them. The game is played on a vertical plane, and the bird moves up and down when you tap the screen. The pipes move from left to right, and the goal of the game is to fly the bird through the pipes without hitting them. The game gets faster as you progress, making it more challenging to keep the bird alive.

## Applications

* **Text Editor:** A simple text editor that you can use to create and edit text files. The text editor supports basic text editing operations, such as cut, copy, paste, and undo. It also has a syntax highlighting feature that can be used to highlight different types of text, such as code, comments, and strings.
* **Calculator:** A simple calculator that you can use to perform basic mathematical operations, such as addition, subtraction, multiplication, and division. The calculator also supports more advanced operations, such as exponentiation and logarithms.
* **To-Do List:** A to-do list application that you can use to keep track of your tasks. The to-do list application allows you to create new tasks, edit existing tasks, and mark tasks as completed. It also allows you to view your tasks by priority, due date, or status.

## Installation

To install this project, you will need to have Python installed on your computer. Once you have Python installed, you can clone the project repository from GitHub.

```
git clone https://github.com/your_username/python-game-and-application-program.git
```

This will create a new directory called `python-game-and-application-program` in your current directory.

## Usage

To run a game or application, you can use the following command:

```
python game_or_application_name.py
```

For example, to run the tic-tac-toe game, you would use the following command:

```
python tic-tac-toe.py
```

## Documentation

The documentation for this project is available in the `docs` directory. The documentation includes instructions on how to install and use the games and applications, as well as a description of the code.

## Issue Tracker

If you find any bugs in this project, please report them in the issue tracker on GitHub. The issue tracker is a great way to communicate with the project developers and to get help with problems that you are having.

## Contribution Guidelines

If you would like to contribute to this project, please read the contribution guidelines in the `CONTRIBUTING.md` file. The contribution guidelines explain how to submit changes to the project, and they also provide a code of conduct that all contributors must follow.